fix: добавить ссылку на трейлер в картинку фильмов. Удалить мусор и часть комментов. Переименовать хук useState
