+++ .eslintrc.json
@@ -11,6 +11,7 @@
     "rules": {
       "react/prop-types": [0, { }],
       "no-underscore-dangle": ["error", { "allow": ["_id", "_checkResponse"] }],
-      "react/jsx-props-no-spreading": "off"
+      "react/jsx-props-no-spreading": "off",
+      "jsx-a11y/label-has-associated-control": "off"
     }
 }
