+++ src/components/App/App.jsx
@@ -1,5 +1,3 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable no-debugger */
 import React, { useEffect, useState } from 'react'; // useRef
 import { Route, Routes, useNavigate } from 'react-router-dom';
 import {
@@ -16,7 +14,7 @@ import NotFound from '../NotFound/NotFound';
 import Footer from '../Footer/Footer';
 import Movies from '../Movies/Movies';
 import SavedMovies from '../SavedMovies/SavedMovies';
-import { useLocalStorageState as useStorage } from '../../utils/hooks';
+import useStorage from '../../utils/hooks';
 import LS_KEYS from '../../constants/localStorageKeys';
 import { setToken } from '../../utils/token'; // getToken, // removeToken,
 import CurrentUserContext from '../../contexts/CurrentUserContext';
@@ -93,17 +91,8 @@ function App() {
     navigate('/', { replace: false });
   };
 
-  // const [clientWidth, setClientWidth] = useState(document.documentElement.clientWidth);
-
-  // useEffect(() => {
-  //   function handleWindowResize() { setClientWidth(document.documentElement.clientWidth); }
-  //   window.addEventListener('resize', handleWindowResize);
-  //   return () => { window.removeEventListener('resize', handleWindowResize); };
-  // }, []);
-
   return (
     <CurrentUserContext.Provider value={currentUserState}>
-      {/* <h2>{clientWidth}</h2> */}
       <Header urlWithHeader={urlWithHeader} isLoggedIn={isLoggedIn} onMenuClick={cbMenuClick} />
 
       <Routes>
