+++ src/components/MoreBtn/MoreBtn.jsx
@@ -13,7 +13,6 @@ function MoreBtn(props) {
         Ещё
       </button>
     </div>
-
   );
 }
