+++ src/components/MoviesCardList/MoviesCardList.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import './MoviesCardList.css';
 import MoviesCard from '../MoviesCard/MoviesCard';
-// import MoreBtn from '../MoreBtn/MoreBtn';
 
 function MoviesCardList(props) {
   // 🟢 это нормально что не все пропсы используются.
@@ -13,18 +12,15 @@ function MoviesCardList(props) {
   } = props;
 
   return (
-    <>
-      <div className="moviesCardList">
-        {filteredMovies.map((movie) => (
-          <MoviesCard
-            key={movie.movieId}
-            movie={movie}
-            setFilteredMovies={setFilteredMovies}
-          />
-        ))}
-      </div>
-      {/* <MoreBtn /> */}
-    </>
+    <div className="moviesCardList">
+      {filteredMovies.map((movie) => (
+        <MoviesCard
+          key={movie.movieId}
+          movie={movie}
+          setFilteredMovies={setFilteredMovies}
+        />
+      ))}
+    </div>
   );
 }
