+++ src/utils/compareStr.js
@@ -1,0 +1,20 @@
+/*   escapeRegExp - Функция для экранирования спец символов в строке,
+которая будет использоваться в регулярке. Например слеш в строке "24/7" или "WTF?".
+Чтобы использовать произвольную строку в качестве части регулярки, нужно убедиться,
+что спец символы регулярок в этой строке воспринимаются движком БУКВАЛЬНО,
+а не как часть синтаксиса регулярки.   */
+function escapeRegExp(string) {
+  // возвращает строку с экранированными спец символами
+  if (typeof string !== 'string') {
+    console.log('НЕ строковой тип данных на входе ');
+    return '';
+  }
+  return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& означает всю найденную строку
+}
+
+// сравниватель строк. 1я строка - запрос. 2я строка - регулярка
+export default function compareStr(str1, str2) {
+  const escapedStr1 = escapeRegExp(str1);
+  const regex = new RegExp(`\\s*${escapedStr1}\\s*`, 'i');
+  return regex.test(str2);
+}
